+++ lib/strong_migrations/checker.rb
@@ -265,8 +265,7 @@ Then add the foreign key in separate migrations."
       # outdated statistics + a new index can hurt performance of existing queries
       if StrongMigrations.auto_analyze && direction == :up && method == :add_index
         if postgresql?
-          # TODO remove verbose in 0.7.0
-          connection.execute "ANALYZE VERBOSE #{connection.quote_table_name(args[0].to_s)}"
+          connection.execute "ANALYZE #{connection.quote_table_name(args[0].to_s)}"
         elsif mariadb? || mysql?
           connection.execute "ANALYZE TABLE #{connection.quote_table_name(args[0].to_s)}"
         end